ICheckListAction extracted the nested class CheckAll
